# path expansion endpoint