--- a/src/bindings/node/node.cc
+++ b/src/bindings/node/node.cc
@@ -218,6 +218,12 @@ private:
                               -> std::string { return actor.transit_available(request); });
   }
 
+  Napi::Value Expansion(const Napi::CallbackInfo& info) {
+    return generic_action(info,
+                          [](valhalla::tyr::actor_t& actor, const std::string& request)
+                              -> std::string { return actor.expansion(request); });
+  }
+
   valhalla::tyr::actor_t actor;
 };