--- a/src/thor/bidirectional_astar.cc
+++ b/src/thor/bidirectional_astar.cc
@@ -183,7 +183,8 @@ void BidirectionalAStar::ExpandForward(GraphReader& graphreader,
     *es = {EdgeSet::kTemporary, idx};
 
     // setting this edge as reached
-    TrackExpansion(graphreader, edgeid, "r");
+    if (expansion_callback_)
+      expansion_callback_(graphreader, "bidirectional_astar", edgeid, "r", false);
   }
 
   // Handle transitions - expand from the end node of each transition
@@ -295,7 +296,8 @@ void BidirectionalAStar::ExpandReverse(GraphReader& graphreader,
     *es = {EdgeSet::kTemporary, idx};
 
     // setting this edge as reached, sending the opposing because this is the reverse tree
-    TrackExpansion(graphreader, oppedge, "r");
+    if (expansion_callback_)
+      expansion_callback_(graphreader, "bidirectional_astar", oppedge, "r", false);
   }
 
   // Handle transitions - expand from the end node of each transition
@@ -326,10 +328,6 @@ BidirectionalAStar::GetBestPath(valhalla::Location& origin,
   costing_ = mode_costing[static_cast<uint32_t>(mode_)];
   travel_type_ = costing_->travel_type();
   access_mode_ = costing_->access_mode();
-
-  // Set up the expansion geojson
-  if (auto expansion = expansion_.lock())
-    rapidjson::Pointer("/properties/algorithm").Set(*expansion, "bidirectional_astar");
 
   // Initialize - create adjacency list, edgestatus support, A*, etc.
   PointLL origin_new(origin.path_edges(0).ll().lng(), origin.path_edges(0).ll().lat());
@@ -435,7 +433,8 @@ BidirectionalAStar::GetBestPath(valhalla::Location& origin,
       edgestatus_forward_.Update(fwd_pred.edgeid(), EdgeSet::kPermanent);
 
       // setting this edge as settled
-      TrackExpansion(graphreader, fwd_pred.edgeid(), "s");
+      if (expansion_callback_)
+        expansion_callback_(graphreader, "bidirectional_astar", fwd_pred.edgeid(), "s", false);
 
       // Prune path if predecessor is not a through edge or if the maximum
       // number of upward transitions has been exceeded on this hierarchy level.
@@ -455,7 +454,8 @@ BidirectionalAStar::GetBestPath(valhalla::Location& origin,
       edgestatus_reverse_.Update(rev_pred.edgeid(), EdgeSet::kPermanent);
 
       // setting this edge as settled, sending the opposing because this is the reverse tree
-      TrackExpansion(graphreader, rev_pred.opp_edgeid(), "s");
+      if (expansion_callback_)
+        expansion_callback_(graphreader, "bidirectional_astar", rev_pred.opp_edgeid(), "s", false);
 
       // Prune path if predecessor is not a through edge
       if ((rev_pred.not_thru() && rev_pred.not_thru_pruning()) ||
@@ -517,7 +517,8 @@ bool BidirectionalAStar::SetForwardConnection(GraphReader& graphreader, const BD
   }
 
   // setting this edge as connected
-  TrackExpansion(graphreader, pred.edgeid(), "c");
+  if (expansion_callback_)
+    expansion_callback_(graphreader, "bidirectional_astar", pred.edgeid(), "c", false);
 
   return true;
 }
@@ -563,7 +564,8 @@ bool BidirectionalAStar::SetReverseConnection(GraphReader& graphreader, const BD
   }
 
   // setting this edge as connected, sending the opposing because this is the reverse tree
-  TrackExpansion(graphreader, oppedge, "c");
+  if (expansion_callback_)
+    expansion_callback_(graphreader, "bidirectional_astar", oppedge, "c", false);
 
   return true;
 }
@@ -630,7 +632,8 @@ void BidirectionalAStar::SetOrigin(GraphReader& graphreader, valhalla::Location&
     adjacencylist_forward_->add(idx);
 
     // setting this edge as reached
-    TrackExpansion(graphreader, edgeid, "r");
+    if (expansion_callback_)
+      expansion_callback_(graphreader, "bidirectional_astar", edgeid, "r", false);
 
     // Set the initial not_thru flag to false. There is an issue with not_thru
     // flags on small loops. Set this to false here to override this for now.
@@ -704,7 +707,8 @@ void BidirectionalAStar::SetDestination(GraphReader& graphreader, const valhalla
     adjacencylist_reverse_->add(idx);
 
     // setting this edge as settled, sending the opposing because this is the reverse tree
-    TrackExpansion(graphreader, edgeid, "r");
+    if (expansion_callback_)
+      expansion_callback_(graphreader, "bidirectional_astar", edgeid, "r", false);
 
     // Set the initial not_thru flag to false. There is an issue with not_thru
     // flags on small loops. Set this to false here to override this for now.