this adds a new endpoint to valhalla called `/expansion` which takes a normal `/route` style request but returns geojson. the geojson is a time-series dataset of which edges of the routing graph were touched by the path algorithm (only bidirectional a* implemented so far) in the order they were touched and what type of label was assigned to them.

the geojson is a small as possible (for such a verbose format). it returns only a single feature which has a geometry of type `MultiLineString`. Each `LineString` represents the beginning and end points (so only two coordinates) of each edge the algorithm touched.

the properties of the feature contain two parallel arrays (which match the number of edges/linestrings in the geometry). the first array is a list of the edge_ids and the second array is a list of the statuses. the statuses can be one of `r` (reached), `s` (settled) or `c` (connected).

this should work for multileg routes as well but there is no signifier of when one legs expansion history ends and the next one begins.

there should be little to no performance impact on the `/route` endpoint as the function that tracks this data is both `inline`d and wrapped in an `if` that says don't track this if we werent called from the `/expansion` action/endpont.