@kevinkreiser 2019-06-14
@dnesbitt61 shouldnt be any performance impact, i was very careful about that. i can do performance testing on my own but would there but a specific test suite that you would recommend to verify that?