@dnesbitt61 2019-06-14
I always check Lancaster to San Diego as my long test route, though any long route would work. valhalla_run_route with multi-run option works best (N successive path generation times are averaged to give the result - I usually use 100). I tried to keep a running tally of where this was, but now I have a new machine. I could try before and after with this branch but I have data without hierarchies and shortcuts.