@kevinkreiser 2019-06-14
I noticed a little bit more of a performance penalty locally using `std::weak_ptr` so i changed the whole thing around and decided to pass an `std::function` into the path algorithm that allows you to track what the algorithm is doing from the outside. this is cool for few reasons.

1. its more general. we are using it to do make some geojson but you could do lots of other stuff with it
2. it keeps the logic out of the path algorithm an in the action/api layer
3. its faster. in my tests there was no difference with this code and master:

```
BEFORE:
kkreiser@kcarbon:~/sandbox/valhalla/valhalla/master$ ./valhalla_run_route --config ../valhalla.json --multi-run 100 -j '{"locations":[{"lat":32.697958,"lon":-117.174648,"street":"Park Boulevard"},{"lat":40.042553,"lon":-76.299115}],"costing":"auto"}' | grep -F average
2019/06/14 20:49:13.476930 [INFO] PathAlgorithm GetBestPath average: 120 ms
kkreiser@kcarbon:~/sandbox/valhalla/valhalla/master$ ./valhalla_run_route --config ../valhalla.json --multi-run 100 -j '{"locations":[{"lat":32.697958,"lon":-117.174648,"street":"Park Boulevard"},{"lat":40.042553,"lon":-76.299115}],"costing":"auto"}' | grep -F average
2019/06/14 20:49:27.358655 [INFO] PathAlgorithm GetBestPath average: 119 ms
kkreiser@kcarbon:~/sandbox/valhalla/valhalla/master$ ./valhalla_run_route --config ../valhalla.json --multi-run 100 -j '{"locations":[{"lat":32.697958,"lon":-117.174648,"street":"Park Boulevard"},{"lat":40.042553,"lon":-76.299115}],"costing":"auto"}' | grep -F average
2019/06/14 20:49:41.409925 [INFO] PathAlgorithm GetBestPath average: 119 ms
```

```
AFTER:
kkreiser@kcarbon:~/sandbox/valhalla/valhalla/build$ ./valhalla_run_route --config ../valhalla.json --multi-run 100 -j '{"locations":[{"lat":32.697958,"lon":-117.174648,"street":"Park Boulevard"},{"lat":40.042553,"lon":-76.299115}],"costing":"auto"}' | grep -F average
2019/06/14 20:48:29.718511 [INFO] PathAlgorithm GetBestPath average: 118 ms
kkreiser@kcarbon:~/sandbox/valhalla/valhalla/build$ ./valhalla_run_route --config ../valhalla.json --multi-run 100 -j '{"locations":[{"lat":32.697958,"lon":-117.174648,"street":"Park Boulevard"},{"lat":40.042553,"lon":-76.299115}],"costing":"auto"}' | grep -F average
2019/06/14 20:48:43.827638 [INFO] PathAlgorithm GetBestPath average: 118 ms
kkreiser@kcarbon:~/sandbox/valhalla/valhalla/build$ ./valhalla_run_route --config ../valhalla.json --multi-run 100 -j '{"locations":[{"lat":32.697958,"lon":-117.174648,"street":"Park Boulevard"},{"lat":40.042553,"lon":-76.299115}],"costing":"auto"}' | grep -F average
2019/06/14 20:48:57.987518 [INFO] PathAlgorithm GetBestPath average: 120 ms
```